Make life better for people
Sometimes people who come and visit your website will need to know more about your and your work.
In this case, they naturally tend to visit your GitHub.

But, how do they get there?
Yeah, you guessed it right. By clicking on the button on your website.

So, when this is the best way to reach your GitHub profile, I personally think that URL that is mention here needs to be correct.

I saw this mistake and took it upon myself to improve this situation and I hope that it would help bring smiles on the faces of people from all around the world.